accessible/tests/mochitest/name/test_list.html

<html>

<head>
  <title>nsIAccessible::name calculation for HTML li</title>
  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../name.js"></script>
  <script type="application/javascript"
          src="../events.js"></script>

  <script type="application/javascript">
    async function doTest() {
      // bulletUpdate
      testName("li_end", "1. list end");
      let reordered = waitForEvent(EVENT_REORDER, getNode("list"));
      const li = document.createElement("li");
      li.setAttribute("id", "li_start");
      li.textContent = "list start";
      getNode("list").insertBefore(li, getNode("li_end"));
      await reordered;
      testName("li_start", "1. list start");
      testName("li_end", "2. list end");

      // bulletUpdate2
      reordered = waitForEvent(EVENT_REORDER, getNode("li_end"));
      // change list style type
      const list = getNode("list");
      list.setAttribute("style", "list-style-type: disc;");
      // Flush both the style change and the resulting layout change.
      // Flushing style on its own is not sufficient, because that can
      // leave frames marked with NS_FRAME_IS_DIRTY, which will cause
      // nsTextFrame::GetRenderedText to report the text of a text
      // frame is empty.
      list.offsetWidth; // flush layout (which also flushes style)
      await reordered;
      testName("li_start", kDiscBulletText + "list start");
      testName("li_end", kDiscBulletText + "list end");

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>

</head>

<body>

  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=634200"
     title="crash [@ nsIFrame::StyleVisibility() ]">
    Mozilla Bug 634200
  </a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <ol id="list">
    <li id="li_end">list end</li>
  </ol>

</body>
</html>
